[test] rosbridge v2 protocol testing
advertise/unadvertise operation test cases added (+11).
