+++ package.json
@@ -17,7 +17,7 @@
     "wsserver": "node bin/rosbridge.js",
     "browser": "node test/browser/test-html.js",
     "lint": "eslint --max-warnings=0 index.js lib examples test",
-    "protocol": "mocha test/nodejs/protocol/test-*.js",
+    "protocol": "mocha test/nodejs/protocol/entry.js",
     "ci": "npm run test && npm run browser && npm run protocol"
   },
   "license": "Apache-2.0",
