+++ test/nodejs/protocol/test-advertise.js
@@ -15,27 +15,9 @@
 'use strict';
 
 const assert = require('assert');
-const child = require('child_process');
-const path = require('path');
 const WebSocket = require('ws');
 
-var rosbridge = path.resolve(__dirname, '../../../bin/rosbridge.js');
-
-describe('ROS2 protocol testing: advertise', function() {
-  var webSocketServer;
-  this.timeout(60 * 1000);
-
-  before(function(done) {
-    webSocketServer = child.fork(rosbridge, {silent: true});
-    webSocketServer.stdout.on('data', function(data) {
-      done();
-    });
-  });
-
-  after(function() {
-    webSocketServer.kill('SIGTERM');
-  });
-
+module.exports = function() {
   let testCasesData = [
     {
       title: 'advertise positive case 1',
@@ -112,4 +94,4 @@ describe('ROS2 protocol testing: advertise', function() {
       });
     });
   });
-});
+};
