+++ test/nodejs/protocol/test-unadvertise.js
@@ -15,27 +15,9 @@
 'use strict';
 
 const assert = require('assert');
-const child = require('child_process');
-const path = require('path');
 const WebSocket = require('ws');
 
-var rosbridge = path.resolve(__dirname, '../../../bin/rosbridge.js');
-
-describe('ROS2 protocol testing: unadvertise', function() {
-  var webSocketServer;
-  this.timeout(60 * 1000);
-
-  before(function(done) {
-    webSocketServer = child.fork(rosbridge, {silent: true});
-    webSocketServer.stdout.on('data', function(data) {
-      done();
-    });
-  });
-
-  after(function() {
-    webSocketServer.kill('SIGTERM');
-  });
-
+module.exports = function() {
   let testCasesData = [
     {
       title: 'unadvertise positive case 1',
@@ -96,4 +78,4 @@ describe('ROS2 protocol testing: unadvertise', function() {
       });
     });
   });
-});
+};
